requirements.txt
discord.py
asyncio
pytz
requests
apscheduler
tabulate